Added transactions
